Split splitter package into two
