+++ java/athena/src/test/java/sleeper/athena/metadata/SleeperMetadataHandlerIT.java
@@ -52,7 +52,7 @@ import sleeper.configuration.properties.instance.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
 import sleeper.core.partition.Partition;
 import sleeper.core.statestore.StateStore;
-import sleeper.splitter.SplitPartition;
+import sleeper.splitter.split.SplitPartition;
 import sleeper.statestore.StateStoreFactory;
 
 import java.io.IOException;
