+++ java/splitter/splitter-core/src/main/java/sleeper/splitter/find/FindPartitionsToSplit.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.splitter;
+package sleeper.splitter.find;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
