+++ java/splitter/splitter-core/src/main/java/sleeper/splitter/find/SplitPartitionJobDefinitionSerDe.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.splitter;
+package sleeper.splitter.find;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
