+++ java/splitter/splitter-core/src/test/java/sleeper/splitter/find/SplitPartitionJobDefinitionSerDeTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.splitter;
+package sleeper.splitter.find;
 
 import org.junit.jupiter.api.Test;
