+++ java/system-test/system-test-dsl/src/main/java/sleeper/systemtest/dsl/partitioning/WaitForPartitionSplitting.java
@@ -24,8 +24,8 @@ import sleeper.core.partition.Partition;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.StateStoreException;
 import sleeper.core.util.PollWithRetries;
-import sleeper.splitter.FindPartitionToSplitResult;
-import sleeper.splitter.FindPartitionsToSplit;
+import sleeper.splitter.find.FindPartitionToSplitResult;
+import sleeper.splitter.find.FindPartitionsToSplit;
 import sleeper.systemtest.dsl.instance.SystemTestInstanceContext;
 
 import java.time.Duration;
